+++ client-encryption/src/main/java/io/confluent/kafka/schemaregistry/encryption/FieldEncryptionExecutor.java
@@ -58,7 +58,7 @@ import org.apache.kafka.common.config.ConfigException;
  * encrypted data, use the KEK from KMS to decrypt the DEK, and use the decrypted DEK to decrypt
  * the data.
  */
-public class FieldEncryptionExecutor implements FieldRuleExecutor {
+public class FieldEncryptionExecutor extends FieldRuleExecutor {
 
   public static final String TYPE = "ENCRYPT";
 
@@ -75,16 +75,11 @@ public class FieldEncryptionExecutor implements FieldRuleExecutor {
 
   private Map<DekFormat, Cryptor> cryptors;
   private Map<String, ?> configs;
-  private Boolean preserveSource;
   private int cacheExpirySecs = -1;
   private int cacheSize = 10000;
   private DekRegistryClient client;
 
   public FieldEncryptionExecutor() {
-  }
-
-  public boolean isPreserveSource() {
-    return Boolean.TRUE.equals(preserveSource);
   }
 
   @Override
@@ -94,11 +89,8 @@ public class FieldEncryptionExecutor implements FieldRuleExecutor {
 
   @Override
   public void configure(Map<String, ?> configs) {
+    super.configure(configs);
     this.configs = configs;
-    Object preserveSourceConfig = configs.get(ENCRYPT_PRESERVE_SOURCE);
-    if (preserveSourceConfig != null) {
-      this.preserveSource = Boolean.parseBoolean(preserveSourceConfig.toString());
-    }
     Object cacheExpirySecsConfig = configs.get(CACHE_EXPIRY_SECS);
     if (cacheExpirySecsConfig != null) {
       try {
@@ -135,20 +127,6 @@ public class FieldEncryptionExecutor implements FieldRuleExecutor {
     FieldTransform transform = new FieldEncryptionExecutorTransform();
     transform.init(ctx);
     return transform;
-  }
-
-  @Override
-  public Object preTransformMessage(RuleContext ctx, FieldTransform transform, Object message)
-      throws RuleException {
-    if (isPreserveSource()) {
-      try {
-        // We use the target schema
-        message = ctx.target().copyMessage(message);
-      } catch (IOException e) {
-        throw new RuleException("Could not copy source message", e);
-      }
-    }
-    return message;
   }
 
   private Cryptor getCryptor(RuleContext ctx) {
@@ -225,12 +203,6 @@ public class FieldEncryptionExecutor implements FieldRuleExecutor {
       cryptor = getCryptor(ctx);
       kekName = getKekName(ctx);
       kek = getKek(ctx, kekName);
-      if (FieldEncryptionExecutor.this.preserveSource == null) {
-        String preserveValueConfig = ctx.getParameter(ENCRYPT_PRESERVE_SOURCE);
-        if (preserveValueConfig != null) {
-          FieldEncryptionExecutor.this.preserveSource = Boolean.parseBoolean(preserveValueConfig);
-        }
-      }
     }
 
     protected String getKekName(RuleContext ctx) throws RuleException {
